+++ src/config/dictionary/generics/sources.mjs
@@ -1,3 +1,6 @@
 export const SOURCE_CATEGORIES = {
   legacy: [23, 26],
+  excluded: [
+    3, 8, 9, 11, 13, 28,
+  ],
 };