+++ src/parser/monster/source.js
@@ -3,7 +3,15 @@ import { DDBSources } from "../../lib/_module.mjs";
 import DDBMonster from "../DDBMonster.js";
 
 DDBMonster.prototype._generateSource = function _generateSource() {
-  const ddbSource = CONFIG.DDB.sources.find((cnf) => cnf.id == this.source.sourceId);
+
+  let ddbSource = CONFIG.DDB.sources.find((cnf) => cnf.id == this.source.sourceId);
+  const ddbSources = (this.source.sources ?? []).filter((s) => s.sourceType === 1);
+  if (this.source.sources && ddbSources.length > 1) {
+    const highestSource = ddbSources.reduce((prev, current) => {
+      return prev.sourceId > current.sourceId ? prev : current;
+    });
+    ddbSource = CONFIG.DDB.sources.find((cnf) => cnf.id == highestSource.sourceId);
+  }
 
   const source = {
     book: ddbSource ? ddbSource.name : "Homebrew",
@@ -15,7 +23,9 @@ DDBMonster.prototype._generateSource = function _generateSource() {
 
   DDBSources.tweakSourceData(source);
 
-  this.npc.system.details.source = source;
+  this.npc.system.source = source;
+  foundry.utils.setProperty(this.npc, "flags.ddbimporter.sourceId", ddbSource.id);
+  foundry.utils.setProperty(this.npc, "flags.ddbimporter.sourceCategory", ddbSource.sourceCategoryId);
 
   this.legacy = CONFIG.DDB.sources.some((ddbSource) =>
     DICTIONARY.sourceCategories.legacy.includes(ddbSource.sourceCategoryId),